Savepoint :D

+++ sources/server/server/Customer.cs
@@ -10,11 +10,11 @@ namespace server
         public String Email { get; set; }
         public String Name { get; set; }
         public String Adr { get; set; }
-        public int Tel { get; set; }
+        public Decimal Tel { get; set; }
 
         public Customer() { }
 
-        public Customer(String Email, String Name, String Adr, int Tel)
+        public Customer(String Email, String Name, String Adr, Decimal Tel)
         {
             this.Email = Email;
             this.Name = Name;
